RFC: Scoped tasks with join handle
This is an early prototype of functionality I would benefit from; it enables you to have tasks that are tied to a scope (like crossbeam threads) rather than lifetime 'static

Issues to deal with, and with which I need help:

 1. There's no way to run a task to completion from `drop` - this causes the tests to run forever because on the basic scheduler, the spawned task loses the race with the main task.
 2. I can't allow you to abort a task, because that leaves it in cancelled state, not completed.

I suspect that with guidance on how to run a task to completion in `drop` so that it's dropped by the scheduler, I could resolve both problems - my guess is that I could just swap the future inside the task for a dummy future to let the task finish its lifecycle without owning anything from the scope, but the locking for that looks hairy enough that I want to be sure it's the right way to do it before I dive in.

Help appreciated!